+++ packages/kitchen/src/components/GlobalStyle/index.ts
@@ -182,7 +182,7 @@ const GlobalStyle = createGlobalStyle`
     }
 
     *, *:before, *:after {
-        box-sizing: inherit;
+        box-sizing: border-box;
     }
 
     /* fonts */